refactor(chat): extract markdown code-fence parsing
